Add `show` subcommand for `session`
The `show` subcommand allow the users to read past conversations. Differently from what got introduced in https://github.com/block/goose/pull/2424, this option allows you to check (read-only) the previous messages from a session without resuming the session.

This is useful for the cases that the user don't want continue a previous session, but rather, only check the content of the conversation and what as executed, without entering the interactive prompt for session.